move typing to another file and remove unnecessary branch in the Result type
